Add files via upload
There are 2 pdf files, one of which describes the problem in a case format. The second pdf file explains the methodologies used and presents a solution to the case. It is formatted as a teaching note and provides sample solutions for the problems mentioned in the case file. The R script is also included which has the regression model and well as data visualization plots to discern the relationships between the different variables.